진짜 마지막 수정..

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: yurlee <yurlee@student.42.kr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/05/04 17:04:00 by yurlee            #+#    #+#             */
-/*   Updated: 2021/05/14 17:53:01 by yurlee           ###   ########.fr       */
+/*   Updated: 2021/05/17 18:03:48 by yurlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ size_t	ft_strlcat(char *dest, const char *src, size_t dstsize)
 	size_t src_size;
 	size_t idx;
 
-	dest_size = (size_t)ft_strlen((char *)dest);
-	src_size = (size_t)ft_strlen((char *)src);
+	dest_size = ft_strlen((char *)dest);
+	src_size = ft_strlen((char *)src);
 	idx = 0;
 	if (dstsize < dest_size + 1)
 		return (dstsize + src_size);
